requirements.txt
pandas==2.2.2
sqlmodel==0.0.16
sqlalchemy==2.0.29
sentence-transformers==2.7.0
faiss-cpu==1.7.4
boto3==1.34.91
botocore==1.34.91
flask==3.0.2
streamlit==1.35.0
openpyxl==3.1.2  # for Excel ingestion if needed
typing-extensions>=4.7.1
